feat: add about list
