+++ styles/pages/about.scss
@@ -47,7 +47,8 @@
 
 @media screen and (max-width: 1300px) {
 	.about-details,
-	.about-anime {
+	.about-anime,
+	.about-list-container {
 		max-width: 100%;
 	}
 }
